+++ Src/HandlerClaim.h
@@ -1,22 +1,23 @@
 #ifndef HANDLER_CLAIM_H
 #define HANDLER_CLAIM_H
 
+#include <cstdint>
 #include <atomic>
 
 struct HandlerClaimState
 {
     public:
         const int counter;
-        const int handlerID;
+        const uint64_t handlerID;
 };
 
 class HandlerClaim
 {
     private:
         std::atomic<int> counter;
-        const int handlerID;
+        const uint64_t handlerID;
     public:
-        HandlerClaim(const int handlerID);
+        HandlerClaim(const uint64_t handlerID);
         HandlerClaim(const HandlerClaim &other);
         HandlerClaimState Claim();
         HandlerClaimState Disclaim();
